fix: circular import for CustomError
