+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './utils';
 export * from './ArweaveWrapper';
+export * from './CustomError';
